[#4310] use <LoadingOverlay /> + fix eslint errors
