+++ akvo/rsr/spa/app/utils/loading-overlay.scss
@@ -1,0 +1,21 @@
+.loading-overlay{
+    position: absolute;
+    height: calc(100vh - 126px);
+    top: 126px;
+    left: 0;
+    width: 100%;
+    background: rgba(255, 255, 255, 0.9);
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    flex-direction: column;
+    z-index: 999999;
+    font-size: 21px;
+    font-weight: 300;
+    .ant-spin{
+      margin-top: 10px;
+    }
+    small{
+      opacity: 0.7;
+    }
+  }
